+++ tests/libc/ansi/stdio/file.c
@@ -29,7 +29,8 @@ debug_file(FILE *f, const char *remarks)
   putchar(f->_flag & 001000  ? 'M' : ' ');
   putchar(f->_flag & 000020  ? 'W' : ' ');
   putchar(f->_flag & 000010  ? 'R' : ' ');
-  printf("  %5d %7p %8p %8d", f->_cnt, f->_ptr, f->_base, f->_bufsiz);
+  printf("  %5d %7p %8p %8lu",
+	 f->_cnt, f->_ptr, f->_base, (unsigned long) f->_bufsiz);
 #endif
   printf("  %s\n", remarks);
 }
@@ -54,7 +55,8 @@ do_mode(const char *rmode, const char *wmode, const char *description)
     setvbuf(f, 0, _IOFBF , 512);
 
 #ifdef __DJGPP__
-    printf("\n---- %s File, Read, bufsiz=%d ----\n", description, f->_bufsiz);
+    printf("\n---- %s File, Read, bufsiz=%lu ----\n",
+	   description, (unsigned long) f->_bufsiz);
     do_headers();
 #endif
 
@@ -122,7 +124,8 @@ do_mode(const char *rmode, const char *wmode, const char *description)
     setvbuf(f, 0, _IOFBF , 512);
 
 #ifdef __DJGPP__
-    printf("\n---- %s File, Write, bufsiz=%d ----\n", description, f->_bufsiz);
+    printf("\n---- %s File, Write, bufsiz=%lu ----\n",
+	   description, (unsigned long) f->_bufsiz);
     do_headers();
 #endif
